requirements.txt
playwright==1.53.0
python-dotenv==1.1.1
kafka-python==2.2.15